APP: Created general Alert Dialog to use across the app

+++ streamflowphone/src/main/java/com/lukakordzaia/streamflowphone/ui/home/ContinueWatchingInfoFragment.kt
@@ -1,8 +1,6 @@
 package com.lukakordzaia.streamflowphone.ui.home
 
 import android.app.Dialog
-import android.graphics.Color
-import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,8 +8,8 @@ import android.view.ViewGroup
 import androidx.navigation.fragment.navArgs
 import com.lukakordzaia.core.utils.AppConstants
 import com.lukakordzaia.core.network.LoadingState
+import com.lukakordzaia.core.utils.DialogUtils
 import com.lukakordzaia.streamflowphone.R
-import com.lukakordzaia.streamflowphone.databinding.DialogRemoveTitleBinding
 import com.lukakordzaia.streamflowphone.databinding.FragmentPhoneContinueWatchingInfoBinding
 import com.lukakordzaia.streamflowphone.ui.baseclasses.BaseBottomSheetVM
 import org.koin.androidx.viewmodel.ext.android.sharedViewModel
@@ -56,30 +54,18 @@ class ContinueWatchingInfoFragment : BaseBottomSheetVM<FragmentPhoneContinueWatc
     private fun fragmentObservers() {
         viewModel.hideContinueWatchingLoader.observe(viewLifecycleOwner, {
             if (it == LoadingState.LOADED) {
-                removeTitleDialog.dismiss()
                 dismiss()
             }
         })
     }
 
     private fun removeTitleDialog() {
-        val binding = DialogRemoveTitleBinding.inflate(LayoutInflater.from(requireContext()))
-        removeTitleDialog = Dialog(requireContext())
-        removeTitleDialog.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
-        removeTitleDialog.setContentView(binding.root)
-
-        if (sharedPreferences.getLoginToken() != "") {
-            binding.title.text = resources.getString(R.string.remove_from_list_title)
-        }
-
-        binding.continueButton.setOnClickListener {
+        removeTitleDialog = DialogUtils.generalAlertDialog(
+            requireContext(),
+            R.string.remove_from_list_title,
+            R.drawable.icon_remove
+        ) {
             viewModel.deleteContinueWatching(args.titleId)
         }
-
-        binding.cancelButton.setOnClickListener {
-            removeTitleDialog.dismiss()
-        }
-
-        removeTitleDialog.show()
     }
 }
